changes to plotter and surrounding classes to get plotter to plot w/ …

+++ arm_controller/chains/urdf_object.py
@@ -31,12 +31,10 @@ class JointType(Enum):
     FLOATING = 'floating'
     PLANAR = 'planar'
 
-
 class URDFJoint:
     """
     Object representation of a URDF Joint
     """
-
     def __init__(self, joint, parent, child, origin, axis=None, limit=None):
         self.name = joint.attrib['name']
         self.type = JointType(joint.attrib['type'])
@@ -45,7 +43,7 @@ class URDFJoint:
         self.origin_xyz = np.array(origin.attrib['xyz'].split()).astype(float)
         self.origin_rpy = np.array(origin.attrib['rpy'].split()).astype(float)
         if axis is not None:
-            self.axis_xyz = np.array(axis.attrib['xyz'].split())
+            self.axis_xyz = np.array(axis.attrib['xyz'].split()).astype(float)
         else:
             self.axis_xyz = None
         if limit is not None:
@@ -70,7 +68,6 @@ class URDFJoint:
             self.limit_upper = None
             self.limit_velocity = None
             self.limit_effort = None
-
 
 class GeometryType(Enum):
     """
@@ -102,7 +99,6 @@ class URDFVisual:
 
         self.mat_name = material.attrib['name']
 
-
 class URDFCollision:
     """
     Object representation of a URDF Collision
@@ -120,7 +116,6 @@ class URDFCollision:
                 self.geometry_attrib[key] = geometry[0].attrib[key]
             else:
                 self.geometry_attrib[key] = float(geometry[0].attrib[key])
-
 
 class URDFLink:
     """
@@ -147,7 +142,6 @@ class URDFLink:
                     coll.find('material')
                 )
             )
-
 
 class URDFObject:
     """
